+++ example/lib/main.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:floating_bottom_navigation_bar/floating_bottom_navigation_bar.dart';
+void main() => runApp(MyApp());
+
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      home: Scaffold(
+        appBar: AppBar(
+          title: Text('Example'),
+        ),
+        //If you want to show body behind the navbar, it should be true
+        extendBody: true,
+        bottomNavigationBar: FloatingNavbar(
+          onTap: (int val) {
+            //returns tab id which is user tapped
+          },
+          currentIndex: 0,
+          items: [
+            FloatingNavbarItem(icon: Icons.home, title: 'Home'),
+            FloatingNavbarItem(icon: Icons.explore, title: 'Explore'),
+            FloatingNavbarItem(icon: Icons.chat_bubble_outline, title: 'Chats'),
+            FloatingNavbarItem(icon: Icons.settings, title: 'Settings'),
+          ],
+        ),
+      ),
+    );
+  }
+}