+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.2.0] - Customization options
+
+* Added borderradius parameters for navbar items, and navbar widget.
+
+
 ## [0.0.1] - TODO: Add release date.
 
 * TODO: Describe initial release.